FIX: Remove pin on hdf5.
conda-forge still pins hdf5 to 1.10.6. 1.12.0 has the MacOS shared libraries so we just add for now the hdf5-static to solve the issue of libraries and comply with the pinned.
If we dont do that we cant install anything that depends on hdf5 since there are no versions compatible with 1.12.0 yet.